More compute separation.
